backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.1.9
langchain-core==0.1.27
langgraph==0.0.26
pydantic==2.6.1
httpx==0.26.0